perf(cache): long-cache issue/PR lists, refresh on sync events
Issue, PR, tracker, and project list queries now use a long stale time so navigating large (500+) lists no longer refetches on every mount. A global sync:completed listener invalidates the matching caches by kind, so data refreshes when a sync runs rather than on a timer.

+++ src/components/layout/sidebar.tsx
@@ -97,7 +97,7 @@ export function AppSidebar({ items, activeId, onSelect }: AppSidebarProps) {
     queryKey: ["projects"],
     queryFn: () => projectService.list(),
     select: (projects) => projects.length,
-    staleTime: cache.staleTime.short,
+    staleTime: cache.staleTime.long,
   });
 
   const itemsBySection = useMemo(() => {
