+++ src/hooks/usePrCards.ts
@@ -38,7 +38,7 @@ export function usePrCards(): UsePrCardsResult {
   const { data: allRepos = [], isLoading: reposLoading } = useQuery({
     queryKey: queryKeys.allRepositories(),
     queryFn: () => repositorySelectionService.listAllSelectedRepositories(),
-    staleTime: cache.staleTime.short,
+    staleTime: cache.staleTime.long,
   });
 
   const prQueries = useQueries({
@@ -51,7 +51,7 @@ export function usePrCards(): UsePrCardsResult {
           scope: resolvePrScope(repo.organization_id, repo.repo_name),
           currentLogin: resolveCurrentLogin(repo.organization_id, organizations, providers),
         }),
-      staleTime: cache.staleTime.short,
+      staleTime: cache.staleTime.long,
     })),
   });
